+++ main.go
@@ -4,6 +4,11 @@ import (
 	"github.com/datacol-io/datacol/cmd"
 )
 
+var (
+	version string
+	rbToken string
+)
+
 func main() {
-	cmd.Initialize()
+	cmd.Initialize(version, rbToken)
 }